Add a very blunt demo how the page could look like.
I'm still working on designing and adding content for the initial page.
Once this is done I'll set it public on an alterantive address for vote
and wait for further input. We also need to discuss which links should
be directly linked to this page using iframe and which should be linked
directly into the confluence. The smaller we keep the page the easier;
but on the other hand if we have nothing at all on the page the
navigation could become a mess for a new user...

Signed-off-by: Andreas Pieber <anpieber@gmail.com>